% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{as.CalDates}
\alias{as.CalDates}
\title{Convert data to class CalDates.}
\usage{
as.CalDates(x)
}
\arguments{
\item{x}{One or more calibrated radiocarbon dates (currently must be of class "BchronCalibratedDates").}
}
\value{
A "CalDates" class object of probabilities per calendar year BP.
}
\description{
Tries to coerce one or more calibrated radiocarbon dates from other formats to an object of class "CalDates" for use by the rcarbon package. Currently only conversion from class "BchronCalibratedDates" in the Bchron package is supported.
}
\examples{
library(Bchron)
onedate <- BchronCalibrate(1300,45,"intcal13")
onedate <- as.CalDates(onedate)
plot(onedate)
}
\seealso{
\code{\link{as.CalGrid}}
}
