% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{as.UncalGrid}
\alias{as.UncalGrid}
\title{Convert data to class UncalGrid.}
\usage{
as.UncalGrid(x)
}
\arguments{
\item{x}{A two-column \code{matrix} or \code{data.frame} class object.}
}
\value{
A CalGrid class object of probabilities or summed probabilities per CRA.
}
\description{
Tries to coerce any two-column matrix or data.frame to a uncalibrated probability distribution (an object of class "UncalGrid") for use by the rcarbon package.
}
\examples{
df <- data.frame(CRA=5000:2000,PrDens=runif(length(5000:2000)))
mycalgrid <- as.UncalGrid(df)
}
