% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barCodes}
\alias{barCodes}
\title{Plot the median values of calibrated radiocarbon dates.}
\usage{
barCodes(x, yrng = c(0, 0.03), width = 20, col = rgb(0, 0, 0, 25,
  maxColorValue = 255), border = NA, fixXorder = FALSE, ...)
}
\arguments{
\item{x}{A \code{CalDates} class object containing calibrated radiocarbon dates.}

\item{yrng}{Number indicating the index value of the calibrated radiocarbon date to be displayed. Default is 1.}

\item{width}{(optional) Character vector to be shown on the top-right corner of the display window.}

\item{fixXorder}{Either \code{'BP'} or \code{'BCAD'}. Indicate whether the calibrated date should be displayed in BP or BC/AD. Default is  \code{'BP'}.}
}
\description{
Plot the median values of multiple calibrated radiocarbon dates in a barcode-like strip.
}
\examples{
to do
}
\seealso{
\code{\link{calibrate}}
}
