% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{BCADtoBP}
\alias{BCADtoBP}
\title{Convert BC/AD dates to BP format}
\usage{
BCADtoBP(x)
}
\arguments{
\item{x}{A numerical vector (currently only basic checks that these numbers are in a sensible range).}
}
\value{
A vector with BP dates.
}
\description{
Converts BC/AD dates to BP format while handling the absence of 'year 0'
}
\examples{
BCADtoBP(-1268)
}
