% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{binPrep}
\alias{binPrep}
\title{Prepare a set of bins for controlling the aggregation of radiocarbon dates
known to be from the same phase of same archaeological site (for use with rspd)}
\usage{
binPrep(sites, ages, h)
}
\arguments{
\item{sites}{a vector of character strings (or number to coerce to character) of all sites or site phases}

\item{h}{a single numeric value passed to hclust() to control degree of grouping of similar ages in a phase site.}

\item{calcurve}{a vector uncalibrated conventional radiocarbon ages}
}
\value{
A vector of character strings of length(ages) that identifying intra-site or intra-phase grouping, for use with rspd()
}
\description{
Used in cases where there is a concern that unusually high levels of sampling for radiocarbon at a given site or in agiven site phase will impede comparison between sites or phases.
}
\examples{
binPrep()
}
