% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{BPtoBCAD}
\alias{BPtoBCAD}
\title{Convert BP dates to BC/AD format}
\usage{
BPtoBCAD(x)
}
\arguments{
\item{x}{A numerical vector (currently only basic checks that these numbers are in a sensible range).}
}
\value{
A vector with BC/BCE dates expressed as negative numbers and AD/CE dates as positive ones.
}
\description{
Converts calibrated BP dates to BC/AD dates, omitting `year 0'
}
\examples{
BPtoBCAD(4200)
}
