% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\alias{calibrate.default}
\title{Calibrate Radiocarbon Dates}
\usage{
calibrate(x, ...)

\method{calibrate}{default}(ages, errors, ids = NA, dateDetails = NA,
  calCurves = "intcal13", resOffsets = 0, resErrors = 0,
  timeRange = c(50000, 0), F14C = FALSE, std = FALSE,
  normalised = FALSE, calMatrix = FALSE, eps = 1e-05, ncores = 1,
  verbose = TRUE)
}
\arguments{
\item{ages}{a vector of radiocarbon ages}

\item{errors}{a vector of errors corresponding to each radiocarbon age}

\item{ids}{a vector of IDs for each date}

\item{dateDetails}{an optional vector of details for each date which will be returned in the output metadata.}

\item{calCurves}{A vector of calibration curves (one between 'intcal13','shcal13' and 'marine13'; default is 'intcal13')}

\item{resOffsets}{A vector of offset values for the Marine Reservoir Effect}

\item{resErrors}{A vector of offset value errors for the Marine Reservoir Effect}

\item{timeRange}{Time range of analysis (in calendar years)}

\item{F14C}{a logical variable indicating whether the calibration should be conducted in F14C space or not. Default is FALSE}

\item{std}{a logical variable indicating whether calibration should be scaled by sqrt(pi*2). Relevant only for F14C space calibration. Default is FALSE.}

\item{normalised}{a logical variable indicating whether the calibration should be normalised or not. Default is FALSE}

\item{calMatrix}{a logical variable indicating whether the age grid should be limited to probabilities higher than \code{eps}}

\item{eps}{Cut-off value for density calculation. Default is 1e-5.}

\item{ncores}{Number of cores used for for parallel execution. Default is 1.}

\item{verbose}{a logical variable indicating whether extra information on progress should be reported. Default is TRUE.}
}
\value{
An object of class CalDates with the following elements
\itemize{
\item{\code{metadata}} {A data.frame containing relevant information regarding each radiocarbon date and the parameter used in the calibration process.}
\item{\code{grids}} {A list of calGrid class objects, containing the posterior probabilities for each calendar year. NA when the parameter calMatrix is set to TRUE} 
\item{\code{calMatrix}} {A matrix of probability values associated with each calendar year between 1 and 50,000 BP (rows) for each radiocarbon date (columns). NA when the parameter calMatrix is set to FALSE}  
}
}
\description{
Function for calibrating radiocarbon dates.
}
\details{
This function computes calibrated radiocarbon ages using the algorithm described in Bronk Ramsey 2008, with the option to process the data in F14C space. Multiple dates can be calibrated using different settings (e.g. calibration curves, reservoir offset, normalisation), and calculations can be executed in parallel to reduce computing time.
}
\examples{
x <- calibrate(ages=4000,errors=30)
plot(x)
# Example with a Marine Date, using a DeltaR of 300 and a DeltaR error of 30
x2 <- calibrate(ages=4000,errors=30,calCurves='marine13',resOffsets=300,resErrors=30)
plot(x)
}
\references{
Bronk Ramsey, C (2008). Radiocarbon dating: Revolutions in Understanding. Archaeometry 50(2), 249–75, DOI: https://doi.org/10.1111/j.1475-4754.2008.00394.x
}
