% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{greatArcDist}
\alias{greatArcDist}
\title{Compute great-arc distances
Function for computing a matrix of great-arc distances (in km) from a set of decimal degree coordinates.}
\usage{
greatArcDist(Latitude, Longitude, verbose = FALSE)
}
\arguments{
\item{Latitude}{A vector lf latitude coordinates in decimal degrees.}

\item{Longitude}{A vector lf longitude coordinates in decimal degrees.}

\item{a}{logical variable indicating whether extra information on progress should be reported. Default is FALSE}
}
\value{
A matrix of great-arc distances in km.
}
\description{
Compute great-arc distances
Function for computing a matrix of great-arc distances (in km) from a set of decimal degree coordinates.
}
\examples{
lon <- c(11.3426,0.1278,0.1218)
lat <- c(44.4949,51.5074,52.2053)
d <- greatArcDist(Latitude=lat,Longitude=lon)
}
