% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{medCal}
\alias{medCal}
\title{Computes the median date of calibrated dates}
\usage{
medCal(x)
}
\arguments{
\item{x}{A \code{CalDates} class object.}
}
\value{
A vector of median dates in cal BP
}
\description{
Function for generating a vector median calibrated dates from a \code{CalDates} class object.
}
\examples{
x <- calibrate(c(3050,2950),c(20,20))
medCal(x)
}
\seealso{
\code{\link{calibrate}}, \code{\link{barCodes}}
}
