% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{modelTest}
\alias{modelTest}
\title{Monte-Carlo simulation test for SPDs}
\usage{
modelTest(x, errors, nsim, bins = NA, runm = NA, timeRange = NA,
  raw = FALSE, model = c("exponential", "explog", "custom"),
  predgrid = NA, calCurves = "intcal13", datenormalised = FALSE,
  spdnormalised = FALSE, ncores = 1, fitonly = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{A vector of radiocarbon ages}

\item{errors}{A vector of errors corresponding to each radiocarbon age}

\item{nsim}{Number of simulations}

\item{bins}{A vector indicating which bin each radiocarbon date is assigned to.}

\item{runm}{A number indicating the window size of the moving average to smooth the SPD. If set to \code{NA} no moving average is applied.Default is NA.}

\item{timeRange}{A vector of length 2 indicating the start and end date of the analysis in cal BP.}

\item{raw}{A logical variable indicating whether all permuted SPDs should be returned or not. Default is FALSE.}

\item{model}{A vector indicating the model to be fitted. One between \code{'exponential'}, \code{'explog'}, and \code{'custom'}.}

\item{predgrid}{A data.frame containing calendar years (column \code{calBP} and associated summed probabilties (column \code{PrDens}). Required when \code{model} is set to \code{'custom'}.}

\item{calCurves}{A vector of calibration curves (one between 'intcal13','shcal13' and 'marine13'; default is 'intcal13')}

\item{datenormalised}{Controls for calibrated dates with probability mass outside the timerange of analysis. If set to TRUE the total probability mass within the time-span of analysis is normalised to sum to unity. Notice that if set to TRUE, calibrations will be normalised (c.f. \code{\link{calibrate}}). Default is FALSE.}

\item{spdnormalised}{A logical variable indicating whether the total probability mass of the SPD is normalised to sum to unity.}

\item{ncores}{Number of cores used for for parallel execution. Default is 1.}

\item{fitonly}{A logical variable. If set to TRUE, only the the model fitting is executed. Default is FALSE.}

\item{verbose}{A logical variable indicating whether extra information on progress should be reported. Default is TRUE.}
}
\value{
An object of class \code{SpdModelTest} with the following elements
\itemize{
\item{\code{result}} {A four column data.frame containing the observed probability density (column \emph{PrDens}) and the lower and the upper values of the simulation envelope (columns \emph{lo} and \emph{hi}) for each calendar year column \emph{calBP}}
\item{\code{sim}} {A matrix containing the simulation results. Available only when \code{raw} is set to TRUE} 
\item{\code{pval}} {A numeric vector containing the p-value of the global signficance test.}  
\item{\code{fit}} {A data.frame containing the probability densities of the fitted model for each calendar year within the time range of analysis}  
\item{\code{coefficients}} {Coefficients of the fitted model. Available only when \code{model} is \code{'exponential'} or \code{'explog'}}  
}
}
\description{
Global and local signficance test for comparing shapes of multiple SPDs using random permutations.
}
\details{
The function implements Timpson et al (2014) Monte-Carlo test for comparing a fitted statistical model to the observed SPD. The null hypothesis is that the observed data are generated by an exponential growth model (fitted using the \code{\link{nls}} function --- implemented with \code{model='exponential'} --- or by log-transforming the probability densities --- implemented with \code{model='explog'}) or by a user defined custom model (supplied as a two-columned data.frame). The fitted model is then 'uncalibrated' (see \code{\link{uncalibratei}}) and \emph{n} samples are randomly drawn in radiocarbon age, with \emph{n} equivalent to the number of bins. The simulated dates are then calibrated and a SPD is generated. This process is repeated \code{nsim} times, so that  a distribution of probabilities densities is assigned to each calendar year. The probabilites are then z-transformed, and a 95% simulation envelope is computed. Local signficance departures are defined as instances where the observed SPD (which is also z-transformed) is outside such envelope. A global signicance is also computed by comparing the total "area" outside the simulation envelope in the observed and simulated data.
}
\references{
Timpson, A., Colledge, S., Crema, E., Edinborough, K., Kerig, T., Manning, K., Thomas, M.G., Shennan, S., (2014). Reconstructing regional population fluctuations in the European Neolithic using radiocarbon dates: a new case-study using an improved method. Journal of Archaeological Science 52, 549–557. doi:10.1016/j.jas.2014.08.011
}
