% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckde.R
\name{plot.compositeKDE}
\alias{plot.compositeKDE}
\title{Plot a composite Kernel Density Estimate of sample dates.}
\usage{
plot.compositeKDE(x, calendar = "BP", type = "envelope", ylim = NA,
  xlim = NA, alpha = NA, fill.col = "lightgrey",
  line.col = "black", line.type = 2, interval = 0.95,
  multiline.col = "black", ...)
}
\arguments{
\item{x}{A \code{compositeKDE} class object generated using the \code{ckde()} function.}

\item{calendar}{Either \code{'BP'} or \code{'BCAD'}. Indicate whether the calibrated date should be displayed in BP or BC/AD. Default is  \code{'BP'}.}

\item{type}{Either \code{envelope} or \code{multiline}. Default is \code{envelope}.}

\item{xlim}{the x limits of the plot. In BP or in BC/AD depending on the choice of the parameter \code{calender}. Notice that if BC/AD is selected BC ages should have a minus sign (e.g. \code{c(-5000,200)} for 5000 BC to 200 AD).}

\item{alpha}{Alpha level for line transparency when \code{type='multiline'}. Default is 10/\code{nsim}, where \code{nsim} is the number of simulations. If \code{nsim} is smaller than 10, \code{alpha} will be set to 1.}

\item{fill.col}{Envelope color when \code{type='envelope'}. Default is 'lightgrey'.}

\item{line.col}{Line color when \code{type='envelope'}. Default is 'black.}

\item{line.type}{Line type when \code{type='envelope'}. Default is 2.}

\item{interval}{Percentile interval of the simulation envelope. Default is 0.95 (i.e. 95%).}

\item{multiline.col}{Line color when \code{type='multiline'}. Default is 'black'.}

\item{...}{Additional arguments affecting the plot}
}
\description{
Plot a composite Kernel Density Estimate of sample dates.
}
\details{
Visualise a \code{compositeKDE} class object. If \code{type} is set \code{'envelope'} an envelope of the percentile iterval defined by the parameter \code{interval} is shown along with the mean KDE. If \code{type} is set \code{'multiline'} all KDEs are shown.
}
\seealso{
\code{\link{ckde}};
}
