% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.spdGG}
\alias{plot.spdGG}
\title{Plot \code{spdGG} class objects}
\usage{
\method{plot}{spdGG}(x, ...)
}
\arguments{
\item{x}{\code{spdGG} class object containing geometric growth rates.}

\item{...}{Additional arguments affecting the plot.}
}
\description{
Plot calibrated geometric growth rates.
}
\seealso{
\code{\link{spd2gg}}
}
