% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.spdRC}
\alias{plot.spdRC}
\title{Plot \code{spdRC} class objects}
\usage{
plot.spdRC(x, calendar = "BP", ...)
}
\arguments{
\item{x}{\code{spdRC} class object containing geometric growth rates.}

\item{calendar}{Either \code{'BP'} or \code{'BCAD'}. Indicate whether the calibrated date should be displayed in BP or BC/AD. Default is  \code{'BP'}.}

\item{...}{Additional arguments affecting the plot.}
}
\description{
Plot rates of change between time-blocks
}
\seealso{
\code{\link{spd2rc}}
}
