% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reScale}
\alias{reScale}
\title{Rescale a numeric vector to a specified minimum and maximum}
\usage{
reScale(x, type = "simple", to = c(0, 1), na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector to smooth.}

\item{type}{what kind of rescaling to perform. Current options are 'simple' (default) and 'normal' which produces a z-score and 'custom' for which the 'to' argument must be specified.}

\item{to}{numeric vector of length 2 specifying the minimum and maximum value to perform a linear rescale between (default is 0 and 1)}

\item{na.rm}{Set to TRUE,this removes NAs before rescaling.}
}
\value{
A numeric vector of rescaled values.
}
\description{
Rescale a numeric vector to a specified minimum and maximum.
}
\examples{
reScale(15:200)
}
