% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spd2rc}
\alias{spd2rc}
\title{Computes rates of change from SPDs}
\usage{
spd2rc(spd, breaks, rate = expression((t2/t1)^(1/d) - 1))
}
\arguments{
\item{spd}{Summed Probability Distribution obtained using the \code{\link{spd}} function.}

\item{breaks}{A vector giving the breakpoints between the time-blocks.}

\item{rate}{An expression defining how the rate of change is calculated, where \code{t1} is the summed probability for a focal block, \code{t2} is the summed probability for next block, and \code{d} is the duration of the blocks. Default is a geometric growth rate (i.e \code{expression((t2/t1)^(1/d)-1)}).}
}
\value{
An object of class \code{spdRC} containing the total summed probability for each time-block and the rate of change between abutting blocks.
}
\description{
Function for computing rates of change between abutting user-defined time-blocks.
}
\examples{
\dontrun{
data(emedyd)
caldates <- calibrate(x=emedyd$CRA, errors=emedyd$Error, normalised=FALSE, calMatrix=TRUE)
bins <- binPrep(sites=emedyd$SiteName, ages=emedyd$CRA, h=50)
emedyd.spd <- spd(caldates,bins,timeRange=c(16000,9000))
emedyd.gg <- spd2rc(emedyd.spd,breaks=seq(16000,9000,-1000))
plot(emedyd.gg)
}
}
