% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{SPpermTest-deprecated}
\alias{SPpermTest-deprecated}
\title{Spatial Permutation Test of summed probability distributions.}
\usage{
SPpermTest(calDates, timeRange, bins, locations, breaks, 
spatialweights,rate=expression((t2/t1)^(1/d)-1), nsim=1000, runm=NA,permute="locations",
ncores=1,datenormalised=FALSE,verbose=TRUE,raw=FALSE)
}
\value{
A \code{spatialTest} class object
}
\description{
This function is deprecated. Please use \code{\link{sptest}} instead.
}
\seealso{
\code{\link{rcarbon-deprecated}}
}
\keyword{internal}
