% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{summary.SpdPermTest}
\alias{summary.SpdPermTest}
\title{Summarise a \code{SpdPermTest} class object}
\usage{
\method{summary}{SpdPermTest}(object, ...)
}
\arguments{
\item{object}{A \code{SpdPermTest} class object produced using the \code{link{permTest}} function.}

\item{...}{Ignored}
}
\description{
\code{summary} method for class "\code{SpdPermTest}"
}
\details{
The summary function returns metadata (number of radiocarbon dates, bins, and simulations), the p-value of the global signficance test, and the chronological interval of local positive and negative deviations from the simulation envelope for each set.
}
\seealso{
\code{\link{permTest}}.
}
