% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{uncalibrate}
\alias{uncalibrate}
\title{Uncalibrate (back-calibrate) a radiocarbon date.}
\usage{
uncalibrate(x, ...)
}
\arguments{
\item{calBP}{A vector of uncalibrated radiocarbon ages .}

\item{CRAerrors}{A vector of standard deviations corresponding to each estimated radiocarbon age.}

\item{roundyear}{An optional vector of IDs for each date.}

\item{calCurves}{A string naming a calibration curve already provided with the rcarbon package (currently 'intcal13', 'shcal13' and 'marine13' are possible; default is 'intcal13' and only one can currently be specified for all dates).}

\item{method}{Either the standard rcarbon method or one that replicates Crema et al 2016.}
}
\value{
A data.frame with specifying the original data, the uncalibrated age without the calibration curve error (ccCRA), the calibration curve error at this point in the curve (ccError), a randomised uncalibrated age (rCRA) given both the stated ccError and any further hypothesised instrumental error provided by the CRAerrors argument (rError).
}
\description{
Function for uncalibrating one or more radiocarbon dates.
}
\details{
This function takes one or more calibrated calendars and looks-up the corresponding uncalibrated age, error of the stated calibration curve at that point. It also provides a randomised estimate of the uncalibrate age based on the curve error (and optionally also a hypothetical measurement error.
}
\examples{
# Uncalibrate two calendar dates
uncalibrate(c(3050,2950))
}
