% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{hpdi}
\alias{hpdi}
\title{Computes the highest posterior density interval (HPDI) for calibrated dates}
\usage{
hpdi(x, credMass = 0.95)
}
\arguments{
\item{x}{A \code{CalDates} class object.}

\item{credMass}{Interval probability mass}
}
\value{
A list of matrices with HPDI ranges and associated probabilities in cal BP. Note that the sum of the probability mass will be only approximately equal to the argument \code{credMass}.
}
\description{
Computes the highest posterior density interval (HPDI) for calibrated dates
}
\examples{
x <- calibrate(c(3050,2950),c(20,20))
hpdi(x)
}
\seealso{
\code{\link{calibrate}}
}
