% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.CalGrid}
\alias{plot.CalGrid}
\title{Plot a summed probability distribution}
\usage{
\method{plot}{CalGrid}(x, calendar = "BP", fill.p = "grey50",
  border.p = NA, xlim = NA, ylim = NA, cex.lab = 0.75,
  cex.axis = cex.lab, mar = c(4, 4, 1, 3), ...)
}
\arguments{
\item{x}{A "CalGrid" class object of summed probabilities per calendar year BP.}

\item{calendar}{Either \code{'BP'} or \code{'BCAD'}. Indicate whether the calibrated date should be displayed in BP or BC/AD. Default is  \code{'BP'}.}

\item{fill.p}{Fill colour of the polygon depicting the summed probability distribution.}

\item{border.p}{Border colour of the polygon depicting the summed probability distribution.}

\item{xlim}{Adjust x axis limits.}

\item{ylim}{Adjust y axis limits (otherwise sensible default).}

\item{cex.lab}{Size of label text.}

\item{cex.axis}{Size of axis text.}

\item{mar}{Adjust margins around plot.}
}
\description{
Plot a summed radiocarbon probability distribution.
}
\examples{
data(euroevol)
mycaldates <- calibrate(euroevol[1:10,"C14Age"], euroevol[1:10,"C14SD"], normalised=FALSE)
myspd <- spd(mycaldates, timeRange=c(8000,2000))
plot(myspd)
}
