% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.spdGG}
\alias{plot.spdGG}
\title{Plot \code{spdGG} class objects}
\usage{
\method{plot}{spdGG}(x, calendar = "BP", ...)
}
\arguments{
\item{x}{\code{spdGG} class object containing geometric growth rates.}

\item{calendar}{Either \code{'BP'} or \code{'BCAD'}. Indicate whether the calibrated date should be displayed in BP or BC/AD. Default is  \code{'BP'}.}

\item{...}{Additional arguments affecting the plot.}
}
\description{
Plot calibrated geometric growth rates.
}
\seealso{
\code{\link{spd2gg}}
}
